chore: ember-cli-update --to 3.13.1 (#149)
* chore: ember-cli-update --to 3.13.1

Also mass-upgrade dependencies

* fix: re-enable node polyfills in ember-auto-import

This fixes a test failure that began appearing after upgrading
ember-auto-import, which stopped automatically enabling webpack
polyfills in v1.4.0. See the URLs below for more details:
https://github.com/ef4/ember-auto-import#i-upgraded-my-ember-auto-import-version-and-now-things-dont-import-what-changed
https://github.com/ef4/ember-auto-import/issues/224#issuecomment-503400386
https://webpack.js.org/configuration/node/

This commit also drops `ember-cli-htmlbars-inline-precompile` since
it is now deprecated and no longer needed.

* ci: test on node 8

* ci: increase memory limit to 2GiB

This is to work around what appears to be a problem with
babel/ember-cli-babel/... on node 8 / 10.