fix: bind _resizeEventHandler to component (fix #5)
Using `window.addEventListener` directly with the
`_resizeEventHandler` method causes its `this` to
refer to `window` instead of the PlotlyComponent instance.
Consequently, the call to Ember's runloop scheduling
methods (e.g. debounce) would not have the context they need,
and errors like the one below would get thrown:

```
Uncaught Error: You attempted to schedule an action in a queue (actions) for a method that doesn't exist
    at DeferredActionQueues.schedule (vendor.js:54347)
    at Backburner._scheduleExpiredTimers (vendor.js:55080)
    at Backburner._runExpiredTimers (vendor.js:55053)
```